removed all SDL-related code and fixed all sorts of compilation errors on windows. Currently no rendering, input or sound goes on at all. In future I will use my lowpoly3d git repo for 3d-rendering and input

+++ include/componentmanager.hpp
@@ -77,7 +77,7 @@ class ComponentManager {
         return birdComponents[id] = BirdComponent();
     }
 
-    static constexpr uint sizePerEntity() {
+    static constexpr size_t sizePerEntity() {
         return
             sizeof(MoveComponent) +
             sizeof(RenderComponent) +
